Move unit test for getPermissionFromMap
