+++ github/util_test.go
@@ -19,86 +19,12 @@ package github
 import (
 	"net/http"
 	"net/url"
-	"reflect"
 	"testing"
 
 	"github.com/fluxcd/go-git-providers/gitprovider"
 	"github.com/fluxcd/go-git-providers/validation"
 	"github.com/google/go-github/v32/github"
 )
-
-func Test_getPermissionFromMap(t *testing.T) {
-	tests := []struct {
-		name        string
-		permissions map[string]bool
-		want        *gitprovider.RepositoryPermission
-	}{
-		{
-			name: "pull",
-			permissions: map[string]bool{
-				"pull":     true,
-				"triage":   false,
-				"push":     false,
-				"maintain": false,
-				"admin":    false,
-			},
-			want: gitprovider.RepositoryPermissionVar(gitprovider.RepositoryPermissionPull),
-		},
-		{
-			name: "push",
-			permissions: map[string]bool{
-				"pull":     true,
-				"triage":   false,
-				"push":     true,
-				"maintain": false,
-				"admin":    false,
-			},
-			want: gitprovider.RepositoryPermissionVar(gitprovider.RepositoryPermissionPush),
-		},
-		{
-			name: "admin",
-			permissions: map[string]bool{
-				"pull":     true,
-				"triage":   true,
-				"push":     true,
-				"maintain": true,
-				"admin":    true,
-			},
-			want: gitprovider.RepositoryPermissionVar(gitprovider.RepositoryPermissionAdmin),
-		},
-		{
-			name: "none",
-			permissions: map[string]bool{
-				"pull":     false,
-				"triage":   false,
-				"push":     false,
-				"maintain": false,
-				"admin":    false,
-			},
-			want: nil,
-		},
-		{
-			name: "false data",
-			permissions: map[string]bool{
-				"pull":     false,
-				"triage":   false,
-				"push":     false,
-				"maintain": false,
-				"admin":    false,
-				"invalid":  true,
-			},
-			want: nil,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			gotPermission := getPermissionFromMap(tt.permissions)
-			if !reflect.DeepEqual(gotPermission, tt.want) {
-				t.Errorf("getPermissionFromMap() = %v, want %v", gotPermission, tt.want)
-			}
-		})
-	}
-}
 
 func Test_validateAPIObject(t *testing.T) {
 	tests := []struct {
